[chores] Moved selenium tests to dedicated new test_widget file
